modules/backend-django/requirements.txt
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.9.1
astroid==3.3.11
async-timeout==5.0.1
attrs==25.3.0
black==25.1.0
blessed==1.21.0
click==8.2.1
dill==0.4.0
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
django-picklefield==3.3
django-q2==1.8.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
flake8==7.3.0
gunicorn==23.0.0
inflection==0.5.1
isort==6.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
phpserialize==1.3
platformdirs==4.3.8
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
pylint==3.3.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==3.5.3
referencing==0.36.2
rpds-py==0.27.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.3
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
wcwidth==0.2.13
